# api/fomo-swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Endpoints labeled open do not require authentication. The protected endpoints on the other hand, do require\nauthentication. I'm not sure how long the jwt token should last. I'm thinking we should set the expire on\nthat token to be super short - like 5 minutes (upto an hour maybe?) to minimize the amount of time an\nattacker can use that token. The refresh token will last longer - currently 7 days. If you make a request\nto a protected endpoint using a \"Refresh\" token in your request headers, you will receive a new\nauthorization token (set-authorization) and refresh token (set-refresh) in the response headers when\nyou make a request with an expired authorization token. You MUST replace both tokens in your request headers\nto stay authenticated. The old refresh token gets replaced on the backend therefore, you need to use the\nnew refresh token to remain actively logged in.",
    "title": "FOMO API",
    "version": "0.0.1"
  },
  "host": "stage.fomo.exchange",
  "basePath": "/api",
  "paths": {
...
    "/users/:id": {
      "put": {
        "description": "You can change the user's first, last, or email. Note we need to implement a secure method of\nverifing the user's new email. This has yet to be implemented.",
        "tags": [
          "users"
        ],
        "summary": "updates user info (protected)",
        "operationId": "UpdateUser",
        "parameters": [
          {
            "x-go-name": "First",
            "description": "Optional.",
            "name": "first",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Last",
            "description": "Optional.",
            "name": "last",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "Email",
            "description": "Optional. Note: we need to validate these!",
            "name": "email",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "\"data\" will contain updated user data with \"status\": \"success\"",
            "schema": {
              "$ref": "#/definitions/responseSuccess"
            }
          },
          "400": {
            "description": "message in badrequest should be descriptive with \"status\": \"fail\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "401": {
            "description": "unauthorized user because of incorrect url param with \"status\": \"fail\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "410": {
            "description": "the user-service is unreachable with status \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      }
    },
    "/users/:id/changepassword": {
      "put": {
        "description": "Allows an authenticated user to change their password. The url param is the user's id.",
        "tags": [
          "users"
        ],
        "summary": "change a user's password (protected)",
        "operationId": "ChangePassword",
        "parameters": [
          {
            "x-go-name": "OldPassword",
            "description": "Required.",
            "name": "oldPassword",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "x-go-name": "NewPassword",
            "description": "Required.",
            "name": "newPassword",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the status will be \"success\" with data null.",
            "schema": {
              "$ref": "#/definitions/responseSuccess"
            }
          },
          "400": {
            "description": "you did something wrong here with status \"fail\". Hopefully, the message is descriptive enough.",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "401": {
            "description": "the user ID in url param does not match with status \"fail\".",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "410": {
            "description": "the user-service is unreachable with status \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          },
          "500": {
            "description": "the message will state what the internal server error was with \"status\": \"error\"",
            "schema": {
              "$ref": "#/definitions/responseError"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Device": {
      "type": "object",
      "properties": {
        "deviceID": {
          "type": "string",
          "x-go-name": "DeviceID"
        },
        "deviceToken": {
          "type": "string",
          "x-go-name": "DeviceToken"
        },
        "deviceType": {
          "type": "string",
          "x-go-name": "DeviceType"
        },
        "externalDeviceID": {
          "type": "string",
          "x-go-name": "ExternalDeviceID"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "UserData": {
      "type": "object",
      "properties": {
        "devices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Device"
          },
          "x-go-name": "Devices"
        },
        "user": {
          "$ref": "#/definitions/UserInfo"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "UserInfo": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "x-go-name": "Email"
        },
        "first": {
          "type": "string",
          "x-go-name": "First"
        },
        "last": {
          "type": "string",
          "x-go-name": "Last"
        },
        "userID": {
          "type": "string",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/asciiu/gomo/api/vendor/github.com/asciiu/gomo/user-service/models"
    },
...
    "responseError": {
      "description": "This response may or may not include data encapsulating the user information.",
      "type": "object",
      "title": "A ResponseSuccess will always contain a status of \"successful\".",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "ResponseError",
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    },
...
    "responseSuccess": {
      "type": "object",
      "title": "A ResponseSuccess will always contain a status of \"successful\".",
      "properties": {
        "data": {
          "$ref": "#/definitions/UserData"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "ResponseSuccess",
      "x-go-package": "github.com/asciiu/gomo/api/controllers"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}